requirements.txt
# TTS Engines
edge-tts>=7.2.3          # Microsoft Edge TTS
gTTS>=2.5.0              # Google TTS
piper-tts>=1.3.0         # Piper TTS (offline) - optional
coqui-tts>=0.27          # Coqui TTS - optional

# Audio Processing
pydub>=0.25.1            # Audio manipulation and combining
simpleaudio>=1.0.4       # Audio playback for preview

# Configuration & Data
pyyaml>=6.0              # YAML parsing
pydantic>=2.11.7         # Data validation

# CLI & Utilities
click>=8.1               # CLI interface
rich>=14.1.0             # Beautiful terminal output

# Development & Testing (optional)
pytest>=8.4.2            # Testing framework
pytest-cov>=6.3.0        # Coverage reporting
